feat: lazy load heatmaps
