+++ src/posthog-core.ts
@@ -69,7 +69,7 @@ import { logger } from './utils/logger'
 import { SessionPropsManager } from './session-props'
 import { isLikelyBot } from './utils/blocked-uas'
 import { extendURLParams, request, SUPPORTS_REQUEST } from './request'
-import { Heatmaps } from './heatmaps'
+import { Heatmaps } from './extensions/heatmaps'
 import { ScrollManager } from './scroll-manager'
 import { SimpleEventEmitter } from './utils/simple-event-emitter'
 import { Autocapture } from './autocapture'
